requirements.txt
python-vlc
flask
pyyaml
